dfu_hostcmd: Move dfu install to protocol/
This moves the rest of the DFU update procedure into
protocol/ so that it is usable outside of htool.
